Exception handling

+++ src/main/java/com/softcraft/dscommerce/services/ProductService.java
@@ -3,13 +3,16 @@ package com.softcraft.dscommerce.services;
 import com.softcraft.dscommerce.dto.ProductDTO;
 import com.softcraft.dscommerce.entities.Product;
 import com.softcraft.dscommerce.repositories.ProductRepository;
+import com.softcraft.dscommerce.services.exceptions.DatabaseException;
+import com.softcraft.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -17,10 +20,9 @@ public class ProductService {
     private ProductRepository repository;
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Optional<Product> result = repository.findById(id);
-        Product product = result.get();
-        ProductDTO productDTO = new ProductDTO(product);
-        return productDTO;
+            Product product = repository.findById(id).orElseThrow(
+                    () -> new ResourceNotFoundException("RECURSO NAO ENCONTRADO!"));
+            return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -39,16 +41,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDTOToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDTOToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
 
     private void copyDTOToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
